+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added 12.1 tabards and some missing tabards to Collections
 - Added 12.1 toys to Collections
 - Added 12.1 pets to Collections
+- Added 12.1 mounts to Collections
 
 ## 9.23.0 - August 23, 2026
 ### Added